Complete mobile layout
Changed font size map to allow for fonts on desktop and mobile to be
accurate.

Added javascript to allow mobile nave to be toggled, along with styles
including an overlay to the body and stopping scrolling when mobile nav
is open.

Added the dropdown menus for the mobile nav and javascript to open them
including animations for the sliding down and creating space.